requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
httpx==0.27.2

pandas==2.2.2
numpy==2.0.2

scikit-learn==1.5.1
xgboost==2.1.1
shap==0.46.0
joblib==1.4.2
matplotlib==3.9.0

interpret>=0.6.4

youtube-transcript-api>=1.0.0
mysql-connector-python>=9.0.0


